requirements.txt
flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
gunicorn==20.1.0